:hammer:refact: CategoryView를 expense 앱으로 이동
