+++ budget_management_project/budget/views.py
@@ -1,26 +1,12 @@
 from django.shortcuts import render
 from rest_framework.views import APIView
-from rest_framework.permissions import IsAuthenticated, AllowAny
+from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework import status
 
-
-from budget_management_project.expense.models import Category
 from budget_management_project.budget.models import Budget
-from budget_management_project.budget.serializers import CategorySerializer, BudgetSerializer
-class CategoryView(APIView):
-    '''
-    카테고리 관련 APIView
-    '''
-    permission_classes = [IsAuthenticated] 
-    
-    def get(self, request):
-        categories = Category.objects.all()
-        serializer = CategorySerializer(categories, many=True)
-        return Response({
-            "message": "카테고리 목록 조회 성공",
-            "data": serializer.data
-        }, status=status.HTTP_200_OK)
+from budget_management_project.budget.serializers import BudgetSerializer
+
 
 class BudgetView(APIView):
     '''
