+++ budget_management_project/expense/views.py
@@ -1,9 +1,28 @@
 from django.db import transaction
 from rest_framework import generics, status
 from rest_framework.response import Response
+from rest_framework.views import APIView
+from rest_framework.permissions import IsAuthenticated
 
-from expense.serializers import ExpenseSerializer
-from expense.permissions import IsExpenseOwner
+from budget_management_project.expense.models import Category
+from budget_management_project.budget.serializers import CategorySerializer
+from budget_management_project.expense.serializers import ExpenseSerializer
+from budget_management_project.expense.permissions import IsExpenseOwner
+
+
+class CategoryView(APIView):
+    '''
+    카테고리 관련 APIView
+    '''
+    permission_classes = [IsAuthenticated] 
+    
+    def get(self, request):
+        categories = Category.objects.all()
+        serializer = CategorySerializer(categories, many=True)
+        return Response({
+            "message": "카테고리 목록 조회 성공",
+            "data": serializer.data
+        }, status=status.HTTP_200_OK)
 
 class ExpenseCreateView(generics.CreateAPIView):
     serializer_class = ExpenseSerializer
